Begin impliment FSoP checkers
Beginning the implimentation of the more complicated Forcing State of
Potential check. This more complicated FSoP is the one in which a subfamily
of K cells is identified as having the exact same K number of potentials;
the remaining cells in the family containing this subfamily need to have
these K potentials removed.