Conceptual breathrough?
After messing around, it seems to be, though it is not yet proven, that
subfamilies have distinct "signatures" based on the sequence the piecewise
sum of their potentials makes. If this is the case, and these signatures
follow what appears to be a simple set of rules, then this may be the best
if not only way so far to identify subfamilies and solve the FSoP problem.