Start add debugging
Started the adding of debugging code (to eventually be optionally toggleable
at runtime); also caught and currently resolving bug where cells are
being put into the queue for having an isolated value even though a member
of one of that cells family is in the queue already awaiting assignment of
that number. Likely fix is just performing the potential change to the
cells families at queue push.