IT PRINTS (a) SUDOKU
Even without FSoP checks implimented, it still sucessfully printed a
valid completed sudoku puzzle. It is remotely possible that the FSoP checks
are actually not necessary as my previous iteration didn't work via a
queue and so maybe came into problems from getting too deep in the call
stack (positive checks would immediately assign numbers which would then
call for more checks). Next is to make the RNG seedable to test several
puzzles to check if this is the case (that the FSoP are not needed in
which case I used a lot of time on this project on something that would
turn out to be irrelevant...) or if the puzzle from the default seed was
just a fluke.