closer still
A commit for just one line? Absolutely. Now sucessfully iterates through
the 2-subfamily groupings and 3-subfamily groupings (012, 013,... 078, 123,
124, ... 178, 234,... 678) but doesn't successfully unwind back; at some
point in the unwinding, the starting index is 5 and the subfamily size is 4
which the helper catches as a valid state (technecally it is but its not
suppose to be checked here!) causing it to then iterate there. If I hadn't
spent so much time trying to both figure out how to make a computer
understand how to catch the FSoP, AND looking into the failed group theory
approach, I might have actually finished this :(