More debug code
Also observation: once a cell has been occupied by a number, the size of
its families ought to be reduced, otherwise it is easy to identify as a
subfamily- from the signature method- the remaining cells in the families
which is no bueno. Also still need to figure out how to restart the
recursive loop once a subfamily has been found.